Fix deleting files by always using realpath()
`ewwwio_delete_file()` uses `realpath()` for the file argument but compares the path against the raw values of the directories which may not give a match. By using `realpath()` for the directories we can make sure that the same data gets compared. This behavior is already the case in `ewwwio_is_file()`.