integrate multiple backgrounds in JS parser
Since JS WebP (and all the parsers really) will ignore elements with JSON, the lazy loader will funnel the image array directly to JS WebP when multiple CSS background images are attached to an element. If the image array is altered by JS WebP, the resulting array is dumped into data-bg-webp for use on the front-end.