use separate webp/avif quality parameters
Use the avif/webp query parameters to define the quality for each format rather than trying to force the quality param to do all those things which results in unexpected behavior. For example, reducing the WebP quality to 60 would also change the quality for resizing images to 60, which is likely not desired.