binary updates and cleanup
Updated gifsicle and pngquant to the latest versions. Removed SunOS support as it just takes way too much time to maintain.
Fixed usage tracking image stats and changed it so that free-exec stats are not reported since their usage is tracked on the API.
Cleaned up a bunch of extra cruft from PEL.