rework how converted files are renamed
Improve the core ewww_image_optimizer_unique_filename() function and EWWW_Image->unique_filename() methods to integrate with the core wp_unique_filename() methods, in case another plugin uses the core wp_unique_filename filter to avoid a filename collision. For instance, when using offloaded images with WP Offload Media.
This ensures that auto-convert and the optional conversion methods all get filtered filenames, at least for the full-size image. The thumbs will then follow whatever was done to the full-size unless it is necessary to further avoid a collision with a mismatched thumb.