tweaking WP Offload Media filters for converted images
Make sure the originals are only added to the WPOM filter if they should be re-offloaded. That is, if Delete Originals is unchecked.
Also add the originals to the WPOM queue when downloading remote to local.
Then, moved our delete filter to priority 21 so that our db records still exist by the time WPOM is deleting images.