fix auto-scale issue in gallery block
When the WP gallery block forces a landscape image into a portrait orientation, the auto-scaler chooses a size that is too small, because an "overscaled" image is needed to accommodate the larger height.
This fix checks the expected height against the aspect ratio (if known), and sets a minimum width to compensate.