fixed issue with no GIF thumbs in free API mode
When exec is disabled, and the plugin is operating in free cloud-based mode without an API key, it was still attempting a gifsicle resize and throwing an error as a result (due to missing gifsicle).
This also impacted WP Offload Media, because it won't offload an image until all the thumbs exist.