improved multisite checks in get/set_option wrappers
Only check for multisite + network_activated once, store that as a static property so subsequent get_option/set_option calls are much quicker.
To complete the optimization, the global functions were made into wrappers for the class methods.
This required moving some code from common.php so that it would not run before the class methods were declared and trigger a fatal error.