improvements for webp method
Make sure to disable cloud conversion when key is removed, but only if local conversion is feasible.
Fix function to remove cloud key, so that it can run via the 'none' redirect mode. Specifically, when a failure is detected with an API key.
Disable either method if not available, worst case this defaults to cloud mode.
Make sure the selected mode is available, otherwise the other mode shows as selected.