use getImageProperty and getImageDepth if possible
Moved to IM native methods to detect color type and bit depth of PNG images. This should reduce memory consumption, and works in the rare case where there is not a local file available.
Only the bit depth is somewhat unreliable, but this is only an issue if/when getImageColors is busted, and then we can fall back to the revamped get_png_color_depth method.