improve quantization for gray indexed png images
When an indexed image is gray, ImageMagick has a bad habit of converting it to Grayscale mode, which uses more bytes.
This can be fixed with png:format=png8, unless the image has an alpha channel with a bit depth > 1. That is, simple transparency with 1 bit alpha can be handled by the png8 hack, but anything beyond that gets ugly.
There's no other way to handle that in ImageMagick, but we can use pngquant to reduce the palette on such problematic images.